requirements.txt
streamlit
pandas
folium
streamlit-folium
openai
ruff
python-dotenv
flask
flask-cors
smartcar
